lavfi/vf_stack_vaapi: factor out the common code for stack setting
The common code will be used in QSV based stack filters.
